Separate sound.cfg parser into new file structure
